Corrección del endpoint de estados
