--- src/main/java/tc3005b224/amazonconnectinsights/dto/AgentDTO.java
@@ -1,13 +1,0 @@
-package tc3005b224.amazonconnectinsights.dto;
-
-public class AgentDTO {
-    private String state;
-
-    public String getState() {
-        return state;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-}
